Memcached and Redis added. Lombok for code less

+++ src/main/java/com/opexos/imageuploader/image/ImageRepository.java
@@ -6,18 +6,22 @@ import org.springframework.data.repository.Repository;
 /**
  * Image repository
  */
-public interface ImageRepository extends Repository<ImageData, Integer> {
+public interface ImageRepository extends Repository<ImageData, Long> {
     /**
      * Saves a given entity
      */
     ImageData save(ImageData entity);
 
+    /**
+     * Returns original image by id
+     */
     @Query("select i.original from ImageData i where i.id = ?1")
-    byte[] getOriginal(int id);
+    byte[] getOriginal(long id);
 
+    /**
+     * Returns preview image by id
+     */
     @Query("select i.preview from ImageData i where i.id = ?1")
-    byte[] getPreview(int id);
-
-
+    byte[] getPreview(long id);
 
 }
